requirements.txt
streamlit
langchain
langchain-community
langchain-text-splitters
langchain-google-genai
chromadb
sentence-transformers
pypdf
google-generativeai